+++ src/main/java/com/sellect/server/order/application/v0/OrderServiceV0Before.java
@@ -12,7 +12,7 @@ import com.sellect.server.order.domain.Orders;
 import com.sellect.server.order.repository.OrderItemRepository;
 import com.sellect.server.order.repository.OrdersRepository;
 import com.sellect.server.order.domain.OrderStatus;
-import com.sellect.server.payment.application.PaymentServiceV0;
+import com.sellect.server.payment.application.v0.PaymentServiceV0;
 import com.sellect.server.payment.domain.Payment;
 import com.sellect.server.product.domain.Inventory;
 import com.sellect.server.product.repository.InventoryRepository;
