+++ src/main/java/com/sellect/server/payment/application/v1/PaymentCompensationServiceV1.java
@@ -1,4 +1,4 @@
-package com.sellect.server.payment.application;
+package com.sellect.server.payment.application.v1;
 
 import com.sellect.server.common.exception.CommonException;
 import com.sellect.server.common.exception.enums.BError;
