+++ src/main/java/com/sellect/server/payment/application/v1/PaymentServiceV1.java
@@ -1,4 +1,4 @@
-package com.sellect.server.payment.application;
+package com.sellect.server.payment.application.v1;
 
 import com.github.f4b6a3.tsid.TsidCreator;
 import com.sellect.server.common.exception.CommonException;
