+++ src/main/java/com/sellect/server/payment/event/listener/PaymentEventListener.java
@@ -2,7 +2,7 @@ package com.sellect.server.payment.event.listener;
 
 import com.sellect.server.common.exception.CommonException;
 import com.sellect.server.order.Infrastructure.response.KakaoPayReadyResponse;
-import com.sellect.server.payment.application.PaymentServiceV1;
+import com.sellect.server.payment.application.v1.PaymentServiceV1;
 import com.sellect.server.payment.event.KakaoPayApproveEvent;
 import com.sellect.server.payment.event.KakaoPayReadyEvent;
 import com.sellect.server.payment.event.PaymentApproveFailedEvent;
